Delete ts3.1/ for task-worklet
It's not needed now that 3.1 is the lowest version supported by DT.
