+++ types/task-worklet/index.d.ts
@@ -1,12 +1,6 @@
-// Type definitions for task-worklet 0.1
-// Project: https://github.com/developit/task-worklet, https://github.com/googlechromelabs/task-worklet
-// Definitions by: Karol Majewski <https://github.com/karol-majewski>
-// Definitions: https://github.com/DefinitelyTyped/DefinitelyTyped
-// TypeScript Version: 2.9
-
-declare class TaskQueue {
+declare class TaskQueue<T extends TaskDescriptor = any> {
     constructor(options?: Options);
-    postTask(taskName: string, ...args: any[]): TaskQueue.Task;
+    postTask<U extends T = any>(taskName: U['name'], ...args: Parameters<U>): Task<ReturnType<U>>;
     addModule(moduleURL: string): Promise<void>;
 }
 
@@ -14,20 +8,23 @@ interface Options {
     size?: number;
 }
 
-declare namespace TaskQueue {
-    interface Task<T = any> {
-        id: number;
-        state: State;
-        result: Promise<T extends PromiseLike<infer U> ? U : T>;
-    }
-
-    type State =
-        | 'cancelled'
-        | 'completed'
-        | 'fulfilled'
-        | 'pending'
-        | 'scheduled';
+interface TaskDescriptor {
+    name: string;
+    (...args: any): any;
 }
 
-export = TaskQueue;
+export interface Task<T = unknown> {
+    id: number;
+    state: State;
+    result: Promise<T extends PromiseLike<infer U> ? U : T>;
+}
+
+export type State =
+    | 'cancelled'
+    | 'completed'
+    | 'fulfilled'
+    | 'pending'
+    | 'scheduled';
+
+export default TaskQueue;
 export as namespace TaskQueue;
