+++ types/task-worklet/package.json
@@ -1,13 +1,5 @@
 {
     "private": true,
-    "types": "index",
-    "typesVersions": {
-        ">=3.1.0-0": {
-            "*": [
-                "ts3.1/*"
-            ]
-        }
-    },
     "dependencies": {
         "typescript": ">2.9.0"
     }
