--- types/task-worklet/ts3.1/tsconfig.json
@@ -1,24 +1,0 @@
-{
-    "compilerOptions": {
-        "module": "commonjs",
-        "lib": [
-            "es6",
-            "dom"
-        ],
-        "noImplicitAny": true,
-        "noImplicitThis": true,
-        "strictNullChecks": true,
-        "baseUrl": "../../",
-        "typeRoots": [
-            "../../"
-        ],
-        "types": [],
-        "noEmit": true,
-        "forceConsistentCasingInFileNames": true,
-        "strictFunctionTypes": true
-    },
-    "files": [
-        "index.d.ts",
-        "task-worklet-tests.ts"
-    ]
-}
